requirements.txt
# Python dependencies
fastmcp>=2.0.0
openai>=1.50.0
pymilvus[milvus-lite]>=2.5.0
FlagEmbedding>=1.2.10
sentence-transformers>=3.0.0
torch>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Google Gemini support (optional - for LLM and embeddings)
google-generativeai>=0.8.0

# MCP HTTP client dependencies
mcp>=1.0.0
httpx-sse>=0.4.0
